Fixes issue with not being able to attach gsn_csm_pres_hgt plots using gsn_attach_plots.
The issue is that the pres_hgt plot requires an overlay on a LogLin plot, which has
tickmarks turned off by default. The gsn_attach_plots function needs to be able to
retrieve tickmark information.

The fix is to temporarily turn on tickmarks for the LogLin plot, via
pmTickMarkDisplayMode, but then turn it off again before you leave the
function. If you don't turn it off, then the Y axis labels will be
incorrect for gsn_csm_pres_hgt.

This addresses NCL-1957.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15263 10fe2890-8a40-4696-8b1d-7ccbb19c398e