Clean up the NetCDF4 options in the ESMF regridding.
The user can have the intermediate files written as NetCDF4, which should hopefully now 
allow them to create files with > 4GB variables.

The one issue is that variable attributes have to be converted to character arrays,
because ESMF_RegridWeightGen uses the Fortran 90 interface to read the NetCDF-4 files,
and this interface can't read NetCDF-4 strings yet.

This addresses NCL-1989.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15343 10fe2890-8a40-4696-8b1d-7ccbb19c398e