Added NormalizeWeights option to ESMF_regrid/_gen_weights.
This replaces the old code in ESMF_regrid_with_weights that applied
the "frac_b" quantity to the weights if using the "conserve" method.

In V6.3.0rp1 of ESMF_RegridWeightGen the "--norm_type fracarea" option
was added, which is now done automatically by ESMF_regrid or
ESMF_regrid_gen_weights if using "conserve".

The old code in ESMF_regrid_with_weights is still there and can be
accessed by setting Opt@WgtApplyFracB = False.

Both of these options are unadvertised, and hence could change.
It's possible that there will be other methods added for
normalizing the weights.

Addresses NCL-2019.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15699 10fe2890-8a40-4696-8b1d-7ccbb19c398e