NCL-2094: fix for case where center is NCEP but sub-center id is unknown.
This case was causing the sub-center to get the center name as a value, but it was not being properly allocated,
leading to a crash. Now the unknown subcenter will be handled like other unknown sub-centers: the id will be given
as a numerical value (converted to a string): in this case it's 223.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15805 10fe2890-8a40-4696-8b1d-7ccbb19c398e