NCL-2098: fix for odd seg fault related to the parameters of a user-defined function
The problem was this line (which appears to have been in the code for ever):
-						for(j = 0; i< tmp_var1->var.n_dims; i++) {
+						for(j = 0; j< tmp_var1->var.n_dims; j++) {
(the plus indicates the line after my fix)
It was causing the crash by letting some subsequent code think a variable had a coorindate variable
when in fact it did not. How this has survived in the code for so long is a bit of a mystery.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15829 10fe2890-8a40-4696-8b1d-7ccbb19c398e