Added wrapper for dim_bfband_n (NCL-1933).
Only minor testing has been done on this. Much more is needed because
this routine has three options for return values.

The original test attached was producing incorrect results, so I'm not
sure what the correct results are supposed to be.

Dennis will have to confirm the correct results once this wrapper is
done.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15879 10fe2890-8a40-4696-8b1d-7ccbb19c398e