NCL-2118: fix for overlay of vectors on vertical plot with 2D coordinates.
This actually applies to any overlay that relies on 2D coordiinates not over a map.
The code was preferring the "Spherical" transformation to the "Curvilinear" transformation,
The spherical transformation is better when overlaying on a map, but obviously not
correct in other cases. Among other problems it limits the y-axis coordinates to the range
-90 - +90. With this fix the spherical transformation is not used unless over a map.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15917 10fe2890-8a40-4696-8b1d-7ccbb19c398e