This fixes the "remap" attribute to have the correct interpolation method when neareststod is used.
This addresses NCL-2130. Note: I'm using a new attribute called
ESMF_regrid_methnod that I think the ESMF group added to the weights
file in a newer version of the offline regridder, rather than the old
"map_method" attribute, which only says "Bilinear" even if "patch" or
"neareststod" was used.

This means that the "remap" attribute attached to the regridded
variable is going to be more verbose than it previously was.

For example, instead of:

  remap :    remapped via ESMF_regrid_with_weights: patch
  remap :    remapped via ESMF_regrid_with_weights: conserve

it will now say:

  remap :    remapped via ESMF_regrid_with_weights: Higher-order Patch
  remap :    remapped via ESMF_regrid_with_weights: First-order Conservative




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15933 10fe2890-8a40-4696-8b1d-7ccbb19c398e