Changed the optional attribute names for draw_color_palette.
The current attributes are:

Frame - whether to advance the frame
LabelsOn - whether to draw the labels for each color
LabelStrings - the labels to use for each color box
LabelFontHeight - the font size for each label
Across - whether to draw the color boxes across (the default) or down.

This addresses (and hopefully resolves) NCL-2122.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@15979 10fe2890-8a40-4696-8b1d-7ccbb19c398e