The cyclic point will now not be added if the range of a 1D longitude coord array is > 360.
Before, this function was just printing a warning. Now, it doesn't do anything.

This is to be on the safe side, because if you let the point be added, the plotted
results will likely be wrong.

I'm a little worried if there's a case where a 1D longitude array has
a range > 360 and you *do* want the cyclic point added.

This addresses NCL-1213.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16052 10fe2890-8a40-4696-8b1d-7ccbb19c398e