Added the first cut at a wrapper for writing intermediate WPS files.
This addresses NCL-2164.

Note: this wrapper is going to change, because "nx" and "ny" don't
need to be passed in by the user.  These can be figured out by the
wrapper.  I'm checking in a version that includes the nx and ny
in case we need to revert to it.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16055 10fe2890-8a40-4696-8b1d-7ccbb19c398e