NCL-2236: partial fix for constant fill with AreaFill mode set
This works a bit better than the original hack but it is not yet totally reliable.
It may not work when the data field is bigger than the area displayed within the
viewport (i.e. over a map when the view is limited using map limits rather than by
sub-setting the data). 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16255 10fe2890-8a40-4696-8b1d-7ccbb19c398e