Fixed a core dump and simplified sizes of various work arrays in speidx function.
I decided that the temporary "tmp_spei" array should be "ntim" and not
"acum_ntim" in size.  When it is copied back to the output array,
"spei", then only acum_ntim points are copied.

All the work arrays (etpSeries, balanceSeries, acumSeries,
seasonSeries) are now "ntim" in size, instead of a fixed size which
could easily be exceeded. These arrays were hard-coded 5000 in the
original C code.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16265 10fe2890-8a40-4696-8b1d-7ccbb19c398e