Re: JIRA  NCL-2270
I replaced:

  k = round(random_uniform(-0.499999, N-0.500001, N), 3) 

with

  k = round(random_uniform(-0.499999d0, N-0.500001d0, N), 3) 

---
Dennis Shea added a comment - 02/Sep/15 8:02 AM
Sent by user: 

Hello Dennis, 

Thanks for the suggestions. The first solution: 
       k = round(random_uniform(-0.499999d0, N-0.500001d0, N), 3) 
seems to work, at least I couldn't make it fail while trying a few 
thousand times in a row. So I think it is a rounding issue. 
Actually, I am not sure why you use -0.499999 and -0.500001. The documentation says that the bounds of the real values to be generated are exclusive, so I thought using -0.500000 and -0.500000 instead would be sufficient. In my case it works (only when adding d0 though), but I may miss another aspect of the problem. 

Best regards, 

Quentin Lejeune 


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16298 10fe2890-8a40-4696-8b1d-7ccbb19c398e