Fixed bug in gsn_coordinates where logic for marker colors was incorrect.
This addresses NCL-2283.

There is somewhat convoluted logic for determining whether a marker
color has been set to transparent.  Tried to simplify this logic
so that now there's only two code sections: one for drawing
missing marker locations, and one for drawing non-missing marker
locations.

The one drawback is that this may be slower for the case where both
sets of markers are being drawn in the same color.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16307 10fe2890-8a40-4696-8b1d-7ccbb19c398e