Overhauled the spei wrapper and internal code that it calls.
First, I got rid of the "old" speidx wrapper, and renamed
the "spei_driver" to be "speidx". This function is the
one that takes both precipitation and temperature as input.

Second, I modified the internal SPEI C code to return the missing
values at the beginning of the array instead of the end, to match what
"R" returns.  We can choose later not to do it this way.

Third, I modified the internal SPEI C code to remove extra checks for
missing values, since this code simply is not set up to handle them.
The only missing values that come in to play are the lrun-1 missing values
at the beginning of the array.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16320 10fe2890-8a40-4696-8b1d-7ccbb19c398e