I have moved the (alpha) NCL function [ thornthwaite_driver.ncl ] to shea_util from contributed. This 'kinda' hides the function. There is no doc. This will be removed if NCL-2325 is implemented.
contributed has a 'new' function gradLatLon_cfd which uses 
centered finite differences (CFD) to estimate the meridional and 
zonal gradients of a quantity. This was created to evaluate if the
spherical harmonic 'gradsg' function was working correctly. (see NCL-2329).
It is!!!! Only the inverse functions are not working correctly.  

It has the advantage that it does not require a global grid like
the spherical harmonic funtions (SPF). Of course, it is less accurate
than the "highly accurate" SPF functions. The function documentation
and an example showing the difference between the SPF and CFD functions
will be posted today (1/4/2016).



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16361 10fe2890-8a40-4696-8b1d-7ccbb19c398e