Removed 'calculate_segment_values' from shea_util.ncl and added it to contributed.ncl
This was inevitable at some point.

The documentation which will be checked in shortly clearly (Bold & Rad) states 
that this function only works for the gregorian (standard) calendar. The functio
will print a fatal error message and exit if another calendar is used.

Note: the reason for the failure is that 'cd_convert' does not recognize other calendars.





git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16412 10fe2890-8a40-4696-8b1d-7ccbb19c398e