Adding first set of NUG examples to NCL source code.
This partially addresses NCL-73.

I'm not sure about prefixing each script name with "NUG_". In the user guide
document that Karin wrote, they are prefixed with "NCL_".  Since ".ncl" is the suffix,
however, I thought "NUG_" would be better so a user knows it came from the NCL User Guide.

I still need to decide what to do with the data files.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16435 10fe2890-8a40-4696-8b1d-7ccbb19c398e