This fixes gsn_create_text so that you can pass in an array of text strings and get an array of text objects back.
The documentation:

http://www.ncl.ucar.edu/Document/Graphics/Interfaces/gsn_create_text.shtml

said this was the expected behavior, but it was never completely
implemented to allow this.

This partially addresses NCL-2408, but it does NOT fix it.  NCL-2408
wants right-justified text. I'm pretty sure that NCL doesn't allow you
to have several strings, each on a different line, and to
right-justify them with a single resource setting.

Instead, I believe you have to create three individual text objects or
annotations, and right-justify each of them yourself.


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16444 10fe2890-8a40-4696-8b1d-7ccbb19c398e