NCL-2404: updates to NCL's file opening code.
This check-in improves the handling of files with no extension.
The checking to see if a file is in GRIB format now just looks at
the beginning of the file. However, files with recognized GRIB 
extensions are still fully parsed for possible GRIB records.
Files that have no recognizeable format trigger a warning message
and a missing file value is returned. Some previous segfaults are
eliminated.
 


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16451 10fe2890-8a40-4696-8b1d-7ccbb19c398e