NCL-2406: other fixes related to this ticket
I removed a number of routines from NclGRIB.c. These were
used for the old dimension determining algorithm but are not
needed now. Converting the test file to GRIB2 found a problem 
with the GRIB2 reader: product template 12 was not fullly supported.
I fixed this with some changes to NclGRIB2.c. There are still some
other product templates that need more work.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16491 10fe2890-8a40-4696-8b1d-7ccbb19c398e