Added special code to thornthwaite to handle ntim=1.
This addresses NCL-2336, in which if a scalar time dimension
is encountered, the values are propagated as if ntim=12.

That's because the thornthwaite routines expect ntim
to be divisable by 12.

I implemented this for the internal "R" version of the routine too,
which is unadvertised.

I think my next step is going to be combining the two wrappers for
"thornthwaite" (which calls the Fortran routine) and "thornthwaite_r"
(which calls the R C routine) into just "thorntwaite", but then add an
unadvertised "opt" argument that will allow you to switch between
the two.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16500 10fe2890-8a40-4696-8b1d-7ccbb19c398e