Implements NCL-2126, changes to default map database. Previously the LowRes was the default database. For 6.4.0, we decided that the default should be conditional:  LowRes would be used unless any features other than geophysical were requested, in which case the MediumRes would be used. This was because the MediumRes is more up-to-date in its country boundaries, but we want to avoid a performance hit for global maps that depict only land/sea boundaries.
A new keyword, "Dynamic" was introduced for the mpDataBaseVersion resource, and is now the default setting. Thus in the default case (or if the user explicitly sets mpDataBaseVersion = "Dynamic"), the new behavior is as described above. If anything other than "Dynamic" is specified for mpDataBaseVersion, then that database is used regardless of feature-sets (this is thus a means to force old behavior).


git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16504 10fe2890-8a40-4696-8b1d-7ccbb19c398e