NCL-2372, NCL-2337, NCL-2245, NCL-2214
This check-in addresses issues with HDF5 dimensions as well as
group and variable syntax problems. The main points are:
1) print functions now always use the real HDF5 names for groups and
variables, with spaces and other characters that cannot be used in NCL variable names.
2) you can write group and variable names in two ways: using string references or by
replacing invalid (for NCL variable) characters with underscores. Actually you can do both
at the same time if you like. 
3) Groups specified without the $...$ syntax must always start with a '/' character, even
if they are relative to a non-root group. This is necessary to allow backwards compatibility
for '/' used as a numerical operator without any spaces after a file variable name. 
4) Variables names can be specified without a slash if only the variable name is given. NCL
will look for the variable in the current file or group as well as in any child groups. If
multiple variables have the same name, the first one encountered will be returned. This will be
in the same ordering as the variable names returned by getfilevarnames.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16525 10fe2890-8a40-4696-8b1d-7ccbb19c398e