NCL-2456: fixed SuppressClose issue for NetCDF4 files
This check-in also addresses some issues with integrating recent changes
to the Nio library into PyNIO. Specifically a seg-fault in the "advanced"
version of the HDFEOS5 reader. Also because the code for controlling the
"FileStructure" file option is non-standard and was placed directly into
the setfileoption code in BuiltInFuncs.c, it was not accessible to the
Nio library, which does not include BuiltInFuncs. I placed it into lower-level
code where it is available to PyNIO.




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16534 10fe2890-8a40-4696-8b1d-7ccbb19c398e