NCL-2463: fixes problem of strings in nested groups.
Also fixes problems with opening HDF5 files -- opening in
"w" mode no longer deletes the original file and opening in
"c" mode fails if the file exists.
Fixes a problem with some stat macros that are only visible
if compiling with BSD compatibility turned on -- now other macros
that are visible in all Posix environments are used.
Also the H5 version of "flush" is now called to make sure the file is
up-to-date before closing it.
And now when writing variables the file is checked to make sure the 
variable does not exist before calling the "create" routine. Otherwise,
the "open" routine is used. This change may still be needed for some
of the user-defined types.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16545 10fe2890-8a40-4696-8b1d-7ccbb19c398e