NCL-2465: fix for subscripted aggregation dimension with varying time units
Now the time unit values are properly adjusted with arbitrary subsrcripting of
the aggregated time dimension. Formerly in this situation the values were not
always converted to the correct units.
Note that the time coordinate units attached to the first file in the file list are
always used as the common unit. If other files in the list have earlier times 
than the common unit they will be converted to negative values. They will still be
correct (the correct date string can be obtained using the usual routines), but if the files 
in the list are not in cronological order, the following warning will be emitted:

warning:Aggregated dimension coordinate values are non-monotonic; check aggregated file ordering



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16552 10fe2890-8a40-4696-8b1d-7ccbb19c398e