NCL-2442: updated grib2 code tables
Unfortunately the "not found in code table file" errors
that Dennis reported in this ticket cannot be corrected 
because the entries for that were not found are not documented
by the NCEP GRIB2 web pages. Table 4.3 contains the generating process
mapping. The values found in the file, all in the range 62000 to 62016,
probably refer to local "generating processes" related to the NGAC model.
If documentation can be found to describe these I would be happy to add 
them to the tables. The other option to get rid of these admittedly annoying
messages might be to just not print the warnings for this piece of metadata
since it is not usually of that much interest to the data consumer. I would
be happy to consider this.
In the meantime I updated all the NCEP tables with new parameters and one new
table that have been added since he last time I did an update (late 2014).




git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16574 10fe2890-8a40-4696-8b1d-7ccbb19c398e