NCL-1809: Fixed bug with gsn_csm_pres_hgt_vector not recognizing pmLabelBar{Width,Height}F resources.
Rick originally looked into this problem, and determined that the pmLabelBar resources were not 
being passed on to the contour plot. I applied Rick's suggested fix, which did cause the labelbar to move up
slightly, but this is because it was resized. 

It is incumbent on the user to set pmLabelBarOrthogonalPosF to move the labelbar back down, so I don't
consider this a bug.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16588 10fe2890-8a40-4696-8b1d-7ccbb19c398e