NCL-2433: fix for missing contour fill on left side of plot
This is caused by the grid boundary line added to fill
plots by CPCLAM and CTCLAM. In the troublesome cases the 
grid edge separates two visible parts of the rendered data.
Now the grid boundary line is added only if the area 
identifier associated with the grid boundary is greater 
than 0. For NCL this will happen only if cnGridBoundFillColor 
is set to a non-transparent color. 



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16599 10fe2890-8a40-4696-8b1d-7ccbb19c398e