Updated comment about GSN_OLD_LABELBAR.
See ticket NCL-509, which is for deciding whether to make the
PlotManager labelbar the default.

I still can't remember why I decided to discard the PlotManager
labelbar. It's certainly not because I'm glutton for punishment.

There was a good reason for it, but this was long ago.  I'm thinking
that it might have been because it was too hard to control the
font heights of the labelbar labels.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16602 10fe2890-8a40-4696-8b1d-7ccbb19c398e