NCL-2433: A better fix for this issue
All changes to conpack and conpackt are backed out
and the fix is implemented at the HLU level including
the addition of an HLU substitute for CPCLAM.
I now think there is indeed a bug but it is in ezmap,
and involves a difference in the "arbitrary" and
the "fast-track" path in performing map transformations.
I believe it is an issue in the fast-track path.
This is a work-around at the HLU level until the 
root of the problem can be figured out.



git-svn-id: https://subversion.ucar.edu/ncldev/ncarg/trunk@16608 10fe2890-8a40-4696-8b1d-7ccbb19c398e