Fixed internal "ref_line_interp" to handle different dimensionalities for X, Y, and ref line (NCL-1567).
This is only for the case of filling above and below a ref line or a
series of ref lines for XY plots.

The bar chart code is slightly different.  It doesn't currently allow
for multiple ref lines with only one curve. It's easy to work-around
this, so I'm not going to fix it for now. No user has complained
about this that I'm aware of.