Updated simple_legend to use percentages from 0-100 as opposed to fractions from 0-1.
In a previous iteration of this function I renamed the resources that specified
a percentage to end in ..Percent, but the units were actually in terms of a fraction (0-1).
I have updated the function so that all inputs are in units of percentage (0-100).
This aligns the *Percent resources in this function with other NCL *Percent resources
in terms of their valid input value range of 0-100.