Enhanced conform and conform_dims to allow for more flexible conforming of arrays that are the same size, or have degenerate dimensions.
These are odd cases, but it's useful in cases where you don't know
apriori if the thing you are conforming is a scalar or an array of the
same size. You won't need a bunch of "if" statements to check for this.

For example, if the dimensions you want to conform to are 3 x 4 x 5,
then your array to be conformed can be any one of these dimensions:

1 x 4 x 5
3 x 1 x 5
3 x 4 x 1
1 x 1 x 5
3 x 1 x 1

[You get the idea...]