Cleaned up the code that adds an X or Y reference line in gsn_csm_xy.
This addresses NCL-2507 and is part of the clean-up effort for NCL-2277.

There's now a single "add_ref_line" procedure that adds the X or Y reference
line, thus getting rid of duplicate code.

I also made the code that gets the color, thicknesses, and patterns more
generic so it can be used for the X or Y reference line.