Added temporary wrappers ceemdan_opt and eemd_opt.
These use the "opt" argument to pass in parameters, rather than
making the parameters separate arguments.

I still have the ceemdan wrapper, which passes in all the parameters
individually.  I'm doing this because Dennis, Adam, and I are
still figuring out the EEMD stuff, and we're not sure yet which
arguments should be optional, and which should be required.