Addresses Jira ticket NCL-2300 - "Memory not getting freed by NCL's "delete" command (with self-assignment)"
The issue was not really with the delete() command, but rather with self-reassignment (either in whole, or reassignment of a slice of the same variable). Because it was a self reassignment case, the memory for the lefthand side (lhs) was needed to construct the memory for the righthand side (rhs). However, it was not getting freed before the lhs variables was updated with pointers for the rhs.

For the case of a self reassignment in total (e.g., X := X), I made this a NO-OP situation. For the case of a reassignment of just a slice, the rhs memory is constructed, the lfh then release, then assignment of rhs to lhs is completed.

I took the opportunity to clean up some compiler warnings, and refactored out the need for the static variable to indicate self reassignment.