Minor cleanup to add_labelbar that adds min/max labels for labelbar.
This is for the case of cnLabelBarEndStyle = "IncludeMinMaxLabels".

There's a bug in NCL (introduced after V6.3.0) in which the
lbLabelStrings are not generated with the proper min/max labels, *if*
the user sets lbLabelAlignment to anything. The cleanup code
prints a helpful warning, and doesn't assume that the labelbar
strings are a specific length.