Added get_enum_value and check_attr_enum_value to make it easier to check predefined resource values.
These functions are useful when you need to do a "getvalues" to
retrieve a resource value that is normally set as a string, but which
getvalues will return as an enumerated integer.

This change involves a slight overhaul of how resource values are
checked. For example, the "if" statement in this code:

   getvalues pobj
     "trGridType" : tr_grid_type
   end getvalues
  if(any(tr_grid_type.eq.(/3,4/))) then

was replaced with:

  if(any(tr_grid_type.eq.get_enum_value((/"Curvilinear","Spherical"/)))) then

"check_attr_two_values" was replaced with "check_attr_enum_value",
which makes it easier to check for both integer and string versions of
a resource.

OLD WAY:
-  if(check_attr_two_values(r,"lbOrientation","vertical",1,True))

NEW WAY:
+  if(check_attr_enum_value(r,"lbOrientation","vertical"))

We also replaced all calls to "lower_case" with "str_lower".