Fixed segfault in addfile() [NCL-2517]
When an undefined environment variable is present in the path provided
to the NCL function addfile(), a segfault occurs. I believe this
behavior was introduced when _NclOpenFile() was refactored to use
alloca(strlen()) instead of using a fixed size buffer.

The underlying function for opening files, _NclOpenFile(), calls the
function _NGResolvePath() to resolve environment variables (as well as
tilde expressions such as "~/foo.nc" or "~hallock/bar.nc"). When
_NGResolvePath() cannot fully resolve a path due to undefined env
variables being present, it returns NULL.

Previously, that NULL pointer was being passed to strlen(), causing the
segfault.

This fix adds a check immediately after calling _NGResolvePath() to see
if it returned NULL, and if NULL it prints a warning message and exits
gracefully.

I also modified the whitespace on several lines in _NclOpenFile() for
consistency. Most lines use tabs, but several lines were using spaces
for indentation.