Decided to overhaul routine that checks for invalid labelbar resource settings to make it more simple.
Basically, if cnLabelBarEndStyle is set to IncludeMinMaxLabels or ExcludeOuterBoxes, then
lbLabelAlignment shouldn't be set. A warning will be issued and the lbLabelAlignment setting
will be removed.

If cnLabelBarEndLabelsOn (deprecated) is set, there are various checks
that will remove this resource and set cnLabelEndEndStyle instead.

Finally, if lbBoxEndCapStyle is any of the Triangle settings, then
cnLabelBarEndStyle must be IncludeOuterBoxes.