Updated the triple2grid wrapper to change internal options for domain and loop.
This is in reference to the bug reported in NCL-1504. The bug was fixed
in the Fortran TRIP2GRD2 routine, but it was then discovered that
TRIP2GRD3 seems to behave erratically (produces different results on
different systems or between optimized and debug versions).

Dennis and I discussed this today in my office and agreed on the
following:

  - LOOP will be forced to be 0 no matter what. This is what causes
    TRIP2GRD2 to be called. TRIP2GRD3 should NOT be called until
    we figure out what the problem is.

  - DOMAIN will now default to 1.0 instead of 0.0, if not set by the user.
    A value of 0.0 was causing some strange results.