Fixed a problem with optimized Intel compiler producing different values for LC GRIB grids.
This is Dave's work and I'm checking it in for him since he's
out-of-town.

Apparently, the signatures of some calls to EZMAP fortran code were
wrong because they did not include the character length argument. They
were still working, Dave muses, only because the length was not needed
in the routines. He also cleaned up the code by making both GetGrids.c
and NclGRIB2.c use the same InitMapTrans function (now renamed to
_NclInitMapTrans since it has global scope).

See NCL-2558 for details.