Fixed code for XY reference lines to not fail with multiple ref lines.
I cleaned up the get_ref_line_options to better check for the case
where you have requested multiple reference lines, but no fill.

Apparently the fill code was being executed anyway, but with a fill
of transparent, so it looked like it was working.  However, it would
fail the minute you had a different number of X and Y reference lines.

This issue was reported by Will Hobbs. See NCL-2578.