Fix Julian year 0 bug in utInvCalendar2_cal()
Related to [NCL-2091]

While testing, we found that repeated utInvCalendar2_cal() calls with
varying input can cause segfaults. Specifically, when the user_unit_str
parameter is initially a valid date string and is then changed to an
invalid string (for example "days since 0000-01-01 00:00:00”, invalid
because there is no year 0 in the Julian calendar) for the next
invocation of utInvCalendar2_cal(), *prev_user_unit_str is freed and
not reallocated; if the function is called a second time consecutively
with the same invalid user_unit_str, it will result in the attempted
freeing of an unallocated pointer (*prev_user_unit_str).