Addresses NCL-2629: "GFLAS buffers apparently affecting whether text gets drawn".
When patches were made to deal with NCL-1667: "thin lines in contour plots using CellFill/AreaFill", a new SOLID_TEXT_FILL fill-type value was added. The legacy drivers had to be modified to look for this and reset back to SOLID_FILL, because they had no way of interpreting the new value. I mistakenly thought that the WISS workstation type did not need to be modified, but that was the source of this bug.