Minor patch for NCL-2685: "negative array index possible in constant-fill contour plot logic".
This bug originally manifest as a situation Mary encountered whereby NCL would crash for a specific script. She discovered that it happened when NCL was compiled against cairo libraries that she built herself, but NCL would run fine when linked against the Mac system-provided version of cairo. I was able to duplicate that behavior on the Mac, although NCL runs fine in either case under Linux.

Investigating the issue however pointed to the function ContourPlot.c::checkareamap().  In rare cases, it was possible for anode->left_id or anode->right_id to be -1 (which has special meaning in an areamap), causing the areamap to be referenced with an index of -2.  A simple check for that case caused the crashing behaviors to go away, and the fix passes a suite of about 5 cases (attached to the ticket) that Dave provided me that exercise this code.