Fix for NCL-2703 "ICON triangular mesh has streaks when cell-filled contours are drawn over a Robinson projection".
See the Jira ticket for details, but short story is the threshold variable "lpeps" in ..../hlu/map_utils.f was wrong declared (by fortran's implicit typing) as type integer when its usage was intended to be of type float. Computation of its value then resulted in truncation (to zero! in the case of this ticket's dataset and script), which then cause the test for whether a pair of vertices wrapped the dateline to be errnoneous in many cases.