Enhance nggcog to allow numeric input and/or output.
nggcog originally only allowed float input and output.

Now you can input any numeric type, and output either float or double.

This partially addresses NCL-2647, which is part of a larger ticket for
overhauling the wrapper functions. This ticket was originally for nggcog
only, and I may create a duplicate ticket for the larger overhaul project.