Removed unnecessary code from nggcog_W()
If any of the 3 input or 2 output variables are double precision, all
calculations are then done using double precision ("type_calc").

In several places in the code, "if(type_calc == NCL_double)" was used;
in some of these cases, the else block (so nothing was type NCL_double)
contained additional checks such as "if(type_olat == NCL_double)",
which by this point should be impossible. I cleaned up several
instances of this.