Support dimsizes >INT_MAX in dim_pqsort [NCL-2678]
Also affects dim_pqsort_n

A second implementation of dpsort.f for large arrays was added as
dpsort_large.f, using INTEGER*8 as necessary. When sorting an array
whose dimension to be sorted is >INT_MAX, the "large" fortran routines
are used; otherwise the original code is used in order to conserve
memory.