Commiting both J. Lee's and A. Brammer's changes to cd_inv_string.
Jared Lee modified this function to add more options and Alan Brammer
cleaned up the code a little.

Note that option "f" has a different meaning in V6.5.0. This is
because in cd_string, "f" is for "small full month", but in
cd_inv_string it was "2-digit fractional second." This was changed in
NCL V6.5.0 so that "f" now also means "small full month", and "r" is a
new option for "2-digit fractional second".